+++ controllers/productsController.ts
@@ -1,6 +1,7 @@
 import { prisma } from "../db.js";
+import { Request, Response } from "express";
 
-export const getProducts = async (req, res) => {
+export const getProducts = async (req: Request, res: Response) => {
   try {
     const products = await prisma.product.findMany();
     return res.status(200).json({ products });
@@ -9,7 +10,7 @@ export const getProducts = async (req, res) => {
   }
 };
 
-export const addProduct = async (req, res) => {
+export const addProduct = async (req: Request, res: Response) => {
   try {
     const { name, price, description, stock } = req.body;
 
@@ -33,13 +34,24 @@ export const addProduct = async (req, res) => {
   }
 };
 
-export const updateProduct = async (req, res) => {
+interface updateBodyProduct {
+  name: string;
+  price: number;
+  description: string;
+  stock: number;
+}
+
+export const updateProduct = async (
+  req: Request<{ id: string }, {}, updateBodyProduct>,
+  res: Response,
+) => {
   try {
     const { name, price, description, stock } = req.body;
+    const id = req.params.id;
 
     const update = await prisma.product.update({
       where: {
-        id: req.params.id,
+        id,
       },
       data: {
         name,
@@ -49,7 +61,7 @@ export const updateProduct = async (req, res) => {
       },
     });
     return res.status(200).json({ success: "Product updated", data: update });
-  } catch (err) {
+  } catch (err: any) {
     if (err.code === "P2025") {
       return res.status(404).json({ error: "Product not found" });
     }
@@ -57,15 +69,20 @@ export const updateProduct = async (req, res) => {
   }
 };
 
-export const deleteProduct = async (req, res) => {
+export const deleteProduct = async (req: Request, res: Response) => {
   try {
+    const id = req.params.id;
+
+    if (typeof id !== "string") {
+      return res.status(400).json({ error: "Invalid Id" });
+    }
     const deleteProduct = await prisma.product.delete({
       where: {
-        id: req.params.id,
+        id,
       },
     });
     return res.status(200).json({ success: "Product deleted successfully" });
-  } catch (err) {
+  } catch (err: any) {
     if (err.code === "P2025") {
       return res.status(404).json({ error: "Product id not found" });
     }
